[CODER] Implemented Device Mesh support for 3D parallelism
- Added ParallelismDimension enum for DP/TP/PP/Expert dimensions
- Added DeviceMeshCoord struct for mesh coordinates
- Added DeviceMesh class with rank/coord conversion and process group management
- Added MeshState helper methods for mesh state queries
- Added TensorParallelMeshContext for mesh-aware TP (stub for future TP layers)
- Added TPMeshLayerFactory for mesh-aware layer creation (stub for future TP layers)
- Added comprehensive unit tests for DeviceMesh and TPMeshContext
- Fixed LoRA IModule ambiguity build errors
- Fixed TPMemoryEstimator type casting error

+++ src/MLFramework/Conversion/TPMemoryEstimator.cs
@@ -23,13 +23,13 @@ public static class TPMemoryEstimator
         estimate.BaseMemoryMB = analysis.TotalMemoryBytes / 1024 / 1024;
 
         // Memory with TP (each rank stores only a fraction)
-        estimate.TPMemoryPerRankMB = (
+        estimate.TPMemoryPerRankMB = (long)(
             analysis.TotalMemoryBytes - analysis.ParallelizableMemoryBytes +
             analysis.ParallelizableMemoryBytes / worldSize
         ) / 1024 / 1024;
 
         // Communication overhead (temporary buffers)
-        estimate.CommunicationOverheadMB = estimate.TPMemoryPerRankMB * 0.1; // 10% overhead
+        estimate.CommunicationOverheadMB = (long)(estimate.TPMemoryPerRankMB * 0.1); // 10% overhead
 
         // Total memory per rank
         estimate.TotalMemoryPerRankMB = estimate.TPMemoryPerRankMB + estimate.CommunicationOverheadMB;
